pyproject.toml
[tool.poetry]
name = "supportgpt"
version = "0.1.0"
description = ""
authors = ["i1i1 <vanyarybin1@live.ru>"]
readme = "README.md"
packages = [{include = "supportgpt"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.7"
langchain = "^0.0.186"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
